+++ client/CHANGELOG.md
@@ -13,3 +13,4 @@ and this project adheres to [Calendar Versioning](https://calver.org/).
 * Added collapsed button and show it in the URL [LANDGRIF-128](https://vizzuality.atlassian.net/browse/LANDGRIF-128)
 * Scenarios list connected with the API [LANDGRIF-182](https://vizzuality.atlassian.net/browse/LANDGRIF-182)
 * Scenario dropdown for edition and remove [LANDGRIF-156](https://vizzuality.atlassian.net/browse/LANDGRIF-156)
+* Filters and sort for scenarios list [LANDGRIF-168](https://vizzuality.atlassian.net/browse/LANDGRIF-168)